refactor: markdown shared, move models, remove utils types
