+++ src/containers/App/styles.scss
@@ -1,2 +1,2 @@
-@use '../../../assets/_normalize.scss';
-@use '../../../assets/_assets.scss';
+@use '../../assets/_normalize.scss';
+@use '../../assets/_assets.scss';
