--- src/containers/episode/components/Markdown/index.js
@@ -1,1 +1,0 @@
-export { default } from './Markdown';
