+++ src/model/Channels.ts
